Allow non-root block device mounts when using external fuse
The internal libfuse has the ability to drop and restore privileges
using setresuid() when ntfs3g is installed with suid root. However,
ChromeOS does not install with suid root, but rather launches ntfs3g as
a regular user with the CAP_SYS_ADMIN capability. This renders the
internal libfuse privilege dropping code ineffective and unnecessary.
Because of this, the safety check to prevent unsafe use of ntfs3g
compiled with external libfuse can be removed, allowing ntfs3g to be
compiled with external libfuse.